requirements2.txt
autopep8==1.5.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==3.0.4
dnspython==2.0.0
et-xmlfile==1.0.1
idna==2.10
jdcal==1.4.1
lxml==4.6.2
numpy==1.19.4
openpyxl==3.0.5
pandas==1.1.5
pycodestyle==2.6.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.4
regex==2020.11.13
requests==2.25.0
six==1.15.0
soupsieve==2.1
SQLAlchemy==1.3.20
stem==1.8.0
toml==0.10.2
urllib3==1.26.2
